+++ SmallPhotos/Data/IUserAccountRepository.cs
@@ -7,7 +7,7 @@ namespace SmallPhotos.Data
     public interface IUserAccountRepository
     {
         Task CreateNewUserAsync(ClaimsPrincipal user);
-        Task<UserAccount> GetAsync(int userAccountId);
+        Task<UserAccount> GetAsync(long userAccountId);
         Task<UserAccount> GetUserAccountAsync(ClaimsPrincipal user);
         Task<UserAccount> GetUserAccountOrNullAsync(ClaimsPrincipal user);
     }
